# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "QuieroKetchup API",
        "description": "Documentación de endpoints públicos de QuieroKetchup.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/orders": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Obtener mis pedidos",
                "operationId": "955bb77a45ae173a485fadb8befed93f",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Order"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Crear un pedido (requiere login)",
                "description": "Requiere token Sanctum (usuario logueado con rol cliente y estado active). Campos obligatorios: restaurant_id, items[].product_id, items[].quantity. Campos opcionales: currency (default PYG si no se envía), metadata, items[].meta, items[].options y options[].meta. En options[] se puede enviar product_option_item_id (recomendado) o bien una opción custom con name y price; esto es posible porque product_option_item_id es nullable en la migración.",
                "operationId": "2123ba58dadb22bd6f2c8081af562230",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "restaurant_id",
                                    "items"
                                ],
                                "properties": {
                                    "restaurant_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "currency": {
                                        "description": "Opcional. Si no se envía se usa PYG (por migración) o se infiere si todos los productos comparten currency.",
                                        "type": "string",
                                        "example": "PYG",
                                        "nullable": true
                                    },
                                    "metadata": {
                                        "description": "Opcional",
                                        "type": "object",
                                        "nullable": true
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "product_id",
                                                "quantity"
                                            ],
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "meta": {
                                                    "description": "Opcional",
                                                    "type": "object",
                                                    "nullable": true
                                                },
                                                "options": {
                                                    "type": "array",
                                                    "items": {
                                                        "description": "Opcional. Enviar product_option_item_id (recomendado) o name/price si es una opción custom.",
                                                        "properties": {
                                                            "product_option_item_id": {
                                                                "type": "integer",
                                                                "example": 5,
                                                                "nullable": true
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Extra queso",
                                                                "nullable": true
                                                            },
                                                            "price": {
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 1000,
                                                                "nullable": true
                                                            },
                                                            "meta": {
                                                                "type": "object",
                                                                "nullable": true
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "minItems": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/orders/{order}": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Obtener detalle de un pedido mío",
                "operationId": "8114ac185cc1bf301d8862617e497098",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "OrderItemOption": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "product_option_item_id": {
                        "type": "integer",
                        "example": 10,
                        "nullable": true
                    },
                    "name": {
                        "type": "string",
                        "example": "Extra queso"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 1000
                    },
                    "meta": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "OrderItem": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "product": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 10
                            },
                            "name": {
                                "type": "string",
                                "example": "Hamburguesa"
                            },
                            "price": {
                                "type": "number",
                                "format": "float",
                                "example": 15000
                            },
                            "currency": {
                                "type": "string",
                                "example": "PYG"
                            }
                        },
                        "type": "object"
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 2
                    },
                    "unit_price": {
                        "type": "number",
                        "format": "float",
                        "example": 15000
                    },
                    "total_price": {
                        "type": "number",
                        "format": "float",
                        "example": 31000
                    },
                    "meta": {
                        "type": "object",
                        "nullable": true
                    },
                    "options": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderItemOption"
                        }
                    }
                },
                "type": "object"
            },
            "Order": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "status": {
                        "type": "string",
                        "example": "pending"
                    },
                    "total_amount": {
                        "type": "number",
                        "format": "float",
                        "example": 31000
                    },
                    "currency": {
                        "type": "string",
                        "example": "PYG"
                    },
                    "metadata": {
                        "type": "object",
                        "nullable": true
                    },
                    "restaurant": {
                        "$ref": "#/components/schemas/Restaurant"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderItem"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Restaurant": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Sucursal Centro"
                    },
                    "slug": {
                        "type": "string",
                        "example": "sucursal-centro"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "phone": {
                        "type": "string",
                        "nullable": true
                    },
                    "email": {
                        "type": "string",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "nullable": true
                    },
                    "settings": {
                        "type": "object",
                        "nullable": true
                    },
                    "meta": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}